factor out hparams for {augmix, colout, cutmix}
