+++ composer/algorithms/__init__.py
@@ -45,21 +45,21 @@ from composer.algorithms.algorithm_registry import get_algorithm_registry as get
 from composer.algorithms.algorithm_registry import list_algorithms as list_algorithms
 from composer.algorithms.alibi import AlibiHparams as AlibiHparams
 from composer.algorithms.augmix import AugMix as AugMix
-from composer.algorithms.augmix import AugMixHparams as AugMixHparams
 from composer.algorithms.blurpool import BlurPool as BlurPool
 from composer.algorithms.blurpool import BlurPoolHparams as BlurPoolHparams
 from composer.algorithms.channels_last import ChannelsLast as ChannelsLast
 from composer.algorithms.channels_last import ChannelsLastHparams as ChannelsLastHparams
 from composer.algorithms.colout import ColOut as ColOut
-from composer.algorithms.colout import ColOutHparams as ColOutHparams
 from composer.algorithms.cutmix import CutMix as CutMix
-from composer.algorithms.cutmix import CutMixHparams as CutMixHparams
 from composer.algorithms.cutout import CutOut as CutOut
 from composer.algorithms.cutout import CutOutHparams as CutOutHparams
 from composer.algorithms.factorize import Factorize as Factorize
 from composer.algorithms.factorize import FactorizeHparams as FactorizeHparams
 from composer.algorithms.ghost_batchnorm import GhostBatchNorm as GhostBatchNorm
 from composer.algorithms.ghost_batchnorm import GhostBatchNormHparams as GhostBatchNormHparams
+from composer.algorithms.hparams import AugMixHparams as AugMixHparams
+from composer.algorithms.hparams import ColOutHparams as ColOutHparams
+from composer.algorithms.hparams import CutMixHparams as CutMixHparams
 from composer.algorithms.label_smoothing import LabelSmoothing as LabelSmoothing
 from composer.algorithms.label_smoothing import LabelSmoothingHparams as LabelSmoothingHparams
 from composer.algorithms.layer_freezing import LayerFreezing as LayerFreezing
