+++ composer/algorithms/algorithm_registry.py
@@ -4,14 +4,16 @@ from typing import Dict, List, Type
 
 from composer.algorithms.algorithm_hparams import AlgorithmHparams
 from composer.algorithms.alibi import AlibiHparams
-from composer.algorithms.augmix import AugMixHparams
 from composer.algorithms.blurpool import BlurPoolHparams
 from composer.algorithms.channels_last import ChannelsLastHparams
-from composer.algorithms.colout import ColOutHparams
-from composer.algorithms.cutmix import CutMixHparams
+# from composer.algorithms.cutmix import CutMixHparams
 from composer.algorithms.cutout import CutOutHparams
 from composer.algorithms.factorize import FactorizeHparams
 from composer.algorithms.ghost_batchnorm import GhostBatchNormHparams
+# from composer.algorithms.hparams import AlibiHparams
+# from composer.algorithms.hparams import BlurPoolHparams
+# from composer.algorithms.hparams import ChannelsLastHparams
+from composer.algorithms.hparams import AugMixHparams, ColOutHparams, CutMixHparams
 from composer.algorithms.label_smoothing import LabelSmoothingHparams
 from composer.algorithms.layer_freezing import LayerFreezingHparams
 from composer.algorithms.mixup import MixUpHparams
@@ -26,6 +28,23 @@ from composer.algorithms.squeeze_excite import SqueezeExciteHparams
 from composer.algorithms.stochastic_depth import StochasticDepthHparams
 from composer.algorithms.swa.hparams import SWAHparams
 from composer.core.algorithm import Algorithm
+
+# from composer.algorithms.hparams import CutOutHparams
+# from composer.algorithms.hparams import FactorizeHparams
+# from composer.algorithms.hparams import GhostBatchNormHparams
+# from composer.algorithms.hparams import LabelSmoothingHparams
+# from composer.algorithms.hparams import LayerFreezingHparams
+# from composer.algorithms.hparams import MixUpHparams
+# from composer.algorithms.hparams import NoOpModelHparams
+# from composer.algorithms.hparams import ProgressiveResizingHparams
+# from composer.algorithms.hparams import RandAugmentHparams
+# from composer.algorithms.hparams import SAMHparams
+# from composer.algorithms.hparams import ScaleScheduleHparams
+# from composer.algorithms.hparams import SelectiveBackpropHparams
+# from composer.algorithms.hparams import SeqLengthWarmupHparams
+# from composer.algorithms.hparams import SqueezeExciteHparams
+# from composer.algorithms.hparams import StochasticDepthHparams
+# from composer.algorithms.hparams import SWAHparams
 
 registry: Dict[str, Type[AlgorithmHparams]] = {
     'blurpool': BlurPoolHparams,
