+++ composer/algorithms/colout/__init__.py
@@ -1,7 +1,6 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 from composer.algorithms.colout.colout import ColOut as ColOut
-from composer.algorithms.colout.colout import ColOutHparams as ColOutHparams
 from composer.algorithms.colout.colout import colout_batch as colout_batch
 from composer.algorithms.colout.colout import colout_image as colout_image
